requirements.txt
SQLAlchemy==2.0.25
fastapi==0.112.0
fastapi-cli==0.0.5
bcrypt==4.2.0
PyJWT==2.9.0
